Omar

+++ lib/core/utils/price_check.dart
@@ -2,7 +2,7 @@ mixin PriceCheck {
   String? checkPrice(String? value) {
     if (value!.isEmpty) {
       return "من فضلك ادخل المبلغ";
-    } else if (!(int.parse(value) > 0)) {
+    } else if (!(double.parse(value) > 0)) {
       return "من فضلك ادخل المبلغ بشكل صحيح";
     }
     return null;
